Made semicolons optional
After using Swift for a while I've decided moving away from required
semicolons is a good move both on comfort and aesthetics. Semicolons
were already something that I was only including a check for for two
reasons:
 1. Historical/familiarity - I started with Java, moved to C, then
    C#, then C++, etc. Even when I used JavaScript I included
    semicolons.
 2. Extra validation - even though I knew it wasn't needed in any
    scenarios (look how easy it was to remove entirely) I still kept
    it as an extra validation check

One down side to this removal has been Ghoul is now getting closer
and closer to just being Swift++ instead of the unique language I
was wanting it to be. I chose `func` because I didn't like Rust's
`fn` or Kotlin's `fun`. When I first saw Swift I liked the general
style and took the argument labels and the subscript syntax
(subscript operators were already planned from C# experience).

In the end, with the changes I've been making Ghoul will be VERY
similar to Swift but the planned differences are:
 1. Namespaces - full C#-style namespace support. WAY better to work
    with than Swift's modules and light years ahead of Rust's `mod`
    in my opinion.
 2. Compile time checks - Ghoul will put as much validation and
    optimization onto the compiler as possible. Ghoul will handle
    some validation through the contracts, it will also handle as
    much as possible when it comes to `const` which will allow us
    to solve some aspects of the program at compile time
 3. Full C++ style templates - Ghoul will allow for `const _: i32`
    template arguments similar to C++
 4. Cross-platform - While the Swift team are trying to allow Swift
    to be usable on platforms besides iOS and macOS (only Ubuntu so
    far), Ghoul will be supported on Windows, Linux, Android, iOS,
    macOS, etc. The other downside to Swift is that some important
    libraries remain vendor-specific. SwiftUI being iOS and macOS
    only.
 5. Fully free & open source - Ghoul will be entirely free and open
    source. None of the compiler or any of the frameworks will be
    closed source. The GUI library I have planned will be completely
    free and open source and will be supported on all platforms Ghoul
    is supported on.