Fixed bug caused by `NUMBER` tokens not having the meta type `VALUE`
For some reason `NUMBER` tokens were keeping the `NIL` meta type
instead of having the `VALUE` meta type (which `CHAR`, `STRING`,
`SYMBOL`, etc. all have). This was not only wrong but caused an
issue with `parseReturnStmt` which wasn't parsing the return
expression when the immediate token after the `return` was a
number.