Changed to use `case` to separate enum const declaration
`enum` now uses `case` to separate enum consts instead of `,`
just to remove the need for `,` in a situation that would be
very similar to `;`. The `,` really isn't needed in this
scenario and `case` makes the syntax look nicer.

We have reached a point where Ghoul will be VERY familiar to
Swift and really easy to pick up if you already know Swift.