Fixed bug with LLVM completely removing if statements
Basically one of the LLVM passes sees the following:
    return i32 12
    return i32 0
And, even though every program I've ever seen executes from top to bottom,
LLVM "optimizes" this into `return i32 0`. I don't know what function pass
it is or why it does this but the only way I found to fix this in an if
statement is when we add a branch back to `merge` we detect if the last
statement was already a branch or return. If it was a branch or return then
we don't add the branch to merge. This will most likely break again with
complex enough source code. At least I would assume, I wouldn't know as I'm
not able to put my mind into a place where the above "optimization" makes
any sense whatsoever. I'm quite excited to replace LLVM with our own
backend written in Ghoul. Partly because that will be fun and partly
because using LLVM is a complete nightmare of barely any documentation and
tons of weird bugs. Maybe we'll switch to GCC before the Ghoul backend.