Fixed semicolon bug for `CaseStmt`
I forgot to also fix the `;` bug we had with `CompoundStmt` for
`CaseStmt`. Now even within a `CaseStmt` if you try to do:

    let i = 0 let j = 12

The parser will error telling your to separate the statements with a `;`
Valid syntax:

    let i = 0; let j = 12

In the future we will have `let i: i32 = 0, j: i32 = 12` but the above is
just an example of how to use multiple `Stmt` nodes on the same line.